대용량 트래픽&데이터 처리 - 6주차 - 주문 재고 차감/복구 분산락 적용

+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/domain/order/OrderInfo.java
@@ -7,7 +7,7 @@ import java.util.Map;
 @Builder
 public record OrderInfo (
         Long userId,
-        Map<Long, Long> productGrp,
+        Map<Long, Integer> productGrp,
         Long couponId,
         Long totPrice
 ) {
